DateInput: add range option
Fixes k-next/starterkit#90

Add a `range`option to the date input and field which lets you either specifiy how many years before and after the current/selected year should be displayed in the select options. Or, when set to `false`, a text field is shown for the year instead of a select field.